pyproject.toml
[tool.poetry]
package-mode = false
name = "thebear-api"
version = "0.1.0"
description = ""
authors = ["carlos1377 <carloseduardoorso1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.13.1"
fastapi = "^0.110.0"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.28.0"
pytest = "^8.1.1"
httpx = "^0.27.0"
pymysql = "^1.1.0"
cryptography = "^42.0.5"
email-validator = "^2.1.1"
passlib = "^1.7.4"
types-passlib = "^1.7.7.20240327"
pytz = "^2024.1"
python-jose = "^3.3.0"
types-python-jose = "^3.3.4.20240106"
types-pytz = "^2024.1.0.20240417"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
